feat: refactored loanSearchUtils
