+++ test/unit/specs/components/Lend/LoanSearch/LoanSearchCheckboxListFilter.spec.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import { render } from '@testing-library/vue';
 import userEvent from '@testing-library/user-event';
 import LoanSearchCheckboxListFilter from '@/components/Lend/LoanSearch/LoanSearchCheckboxListFilter';
-import { getCheckboxLabel } from '@/util/loanSearchUtils';
+import { getCheckboxLabel } from '@/util/loanSearch/filterUtils';
 
 const getOptions = disabled => [...Array(4)].map((_c, i) => ({
 	id: i,
